requirements.txt
openpyxl~=3.1.5
python-telegram-bot~=21.7
pandas~=2.2.3
selenium~=4.26.1
playwright~=1.39.0
playwright[browsers]
python-dotenv~=1.0.0
email-validator~=2.1.0